refactor: CUDA backend use size_t for index

+++ src/backend/cuda/kernels/kernels.cu
@@ -1,9 +1,9 @@
 #include "backend/cuda/kernels/kernels.cuh"
 
-__device__ __forceinline__ int physicalOffsetCUDA(int linear, const TensorLayout& L) {
-    int off = L.offset;
-    for (int d = L.rank - 1; d >= 0; d--) {
-        int coord = linear % L.shape[d];
+__device__ __forceinline__ size_t physicalOffsetCUDA(size_t linear, const TensorLayout& L) {
+    size_t off = L.offset;
+    for (int d = (int)L.rank - 1; d >= 0; d--) {
+        size_t coord = linear % L.shape[d];
         linear /= L.shape[d];
         off += coord * L.strides[d];
     }
@@ -13,15 +13,15 @@ __device__ __forceinline__ int physicalOffsetCUDA(int linear, const TensorLayout
 __global__ void addKernel(
     const float* __restrict__ lhs, const TensorLayout lhsLayout, 
     const float* __restrict__ rhs, const TensorLayout rhsLayout, 
-    float* __restrict__ out, const TensorLayout outLayout, int count) {
+    float* __restrict__ out, const TensorLayout outLayout, size_t count) {
 
-    int i = blockIdx.x * blockDim.x + threadIdx.x;
+    size_t i = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (i >= count) return;
 
-    int lhsIdx = physicalOffsetCUDA(i, lhsLayout);
-    int rhsIdx = physicalOffsetCUDA(i, rhsLayout);
-    int outIdx = physicalOffsetCUDA(i, outLayout);
+    size_t lhsIdx = physicalOffsetCUDA(i, lhsLayout);
+    size_t rhsIdx = physicalOffsetCUDA(i, rhsLayout);
+    size_t outIdx = physicalOffsetCUDA(i, outLayout);
 
     out[outIdx] = lhs[lhsIdx] + rhs[rhsIdx];
 }
@@ -29,15 +29,15 @@ __global__ void addKernel(
 __global__ void subKernel(
     const float* __restrict__ lhs, const TensorLayout lhsLayout,
     const float* __restrict__ rhs, const TensorLayout rhsLayout,
-    float* __restrict__ out, const TensorLayout outLayout, int count) {
+    float* __restrict__ out, const TensorLayout outLayout, size_t count) {
 
-    int i = blockIdx.x * blockDim.x + threadIdx.x;
+    size_t i = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (i >= count) return;
 
-    int lhsIdx = physicalOffsetCUDA(i, lhsLayout);
-    int rhsIdx = physicalOffsetCUDA(i, rhsLayout);
-    int outIdx = physicalOffsetCUDA(i, outLayout);
+    size_t lhsIdx = physicalOffsetCUDA(i, lhsLayout);
+    size_t rhsIdx = physicalOffsetCUDA(i, rhsLayout);
+    size_t outIdx = physicalOffsetCUDA(i, outLayout);
 
     out[outIdx] = lhs[lhsIdx] - rhs[rhsIdx];
 }
@@ -45,15 +45,15 @@ __global__ void subKernel(
 __global__ void mulKernel(
     const float* __restrict__ lhs, const TensorLayout lhsLayout,
     const float* __restrict__ rhs, const TensorLayout rhsLayout,
-    float* __restrict__ out, const TensorLayout outLayout, int count) {
+    float* __restrict__ out, const TensorLayout outLayout, size_t count) {
 
-    int i = blockIdx.x * blockDim.x + threadIdx.x;
+    size_t i = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (i >= count) return;
 
-    int lhsIdx = physicalOffsetCUDA(i, lhsLayout);
-    int rhsIdx = physicalOffsetCUDA(i, rhsLayout);
-    int outIdx = physicalOffsetCUDA(i, outLayout);
+    size_t lhsIdx = physicalOffsetCUDA(i, lhsLayout);
+    size_t rhsIdx = physicalOffsetCUDA(i, rhsLayout);
+    size_t outIdx = physicalOffsetCUDA(i, outLayout);
 
     out[outIdx] = lhs[lhsIdx] * rhs[rhsIdx];
 }
@@ -61,47 +61,47 @@ __global__ void mulKernel(
 __global__ void divKernel(
     const float* __restrict__ lhs, const TensorLayout lhsLayout,
     const float* __restrict__ rhs, const TensorLayout rhsLayout,
-    float* __restrict__ out, const TensorLayout outLayout, int count) {
+    float* __restrict__ out, const TensorLayout outLayout, size_t count) {
 
-    int i = blockIdx.x * blockDim.x + threadIdx.x;
+    size_t i = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (i >= count) return;
 
-    int lhsIdx = physicalOffsetCUDA(i, lhsLayout);
-    int rhsIdx = physicalOffsetCUDA(i, rhsLayout);
-    int outIdx = physicalOffsetCUDA(i, outLayout);
+    size_t lhsIdx = physicalOffsetCUDA(i, lhsLayout);
+    size_t rhsIdx = physicalOffsetCUDA(i, rhsLayout);
+    size_t outIdx = physicalOffsetCUDA(i, outLayout);
 
     out[outIdx] = lhs[lhsIdx] / rhs[rhsIdx];
 }
 
 
-__global__ void fillKernel(float* __restrict__ data, float value, int count) {
-    unsigned int i = blockIdx.x * blockDim.x + threadIdx.x;
+__global__ void fillKernel(float* __restrict__ data, float value, size_t count) {
+    size_t i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < count) data[i] = value;
 }
 
 
 __global__ void setKernel(
     float* __restrict__ dst, const TensorLayout dstLayout,
-    const float* __restrict__ src, const TensorLayout srcLayout, int count) {
+    const float* __restrict__ src, const TensorLayout srcLayout, size_t count) {
 
-    int i = blockIdx.x * blockDim.x + threadIdx.x;
+    size_t i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i >= count) return;
 
-    int dstIdx = physicalOffsetCUDA(i, dstLayout);
-    int srcIdx = physicalOffsetCUDA(i, srcLayout);
+    size_t dstIdx = physicalOffsetCUDA(i, dstLayout);
+    size_t srcIdx = physicalOffsetCUDA(i, srcLayout);
     dst[dstIdx] = src[srcIdx];
 }
 
 __global__ void checkAllEqualKernel(
     const float* __restrict__ lhs, const TensorLayout lhsLayout,
     const float* __restrict__ rhs, const TensorLayout rhsLayout,
-    int* isEqualFlag, int count) {
+    int* isEqualFlag, size_t count) {
 
-    unsigned int i = blockIdx.x * blockDim.x + threadIdx.x;
+    size_t i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i >= count) return;
 
-    int lhsIdx = physicalOffsetCUDA(i, lhsLayout);
-    int rhsIdx = physicalOffsetCUDA(i, rhsLayout);
+    size_t lhsIdx = physicalOffsetCUDA(i, lhsLayout);
+    size_t rhsIdx = physicalOffsetCUDA(i, rhsLayout);
     if (lhs[lhsIdx] != rhs[rhsIdx]) *isEqualFlag = 0;
 }
