+++ src/backend/cuda/kernels/kernels.cuh
@@ -3,38 +3,38 @@
 
 #include "nforge/core/tensor_layout.h"
 
-__device__ __forceinline__ int physicalOffsetCUDA(int linear, const TensorLayout& L);
+__device__ __forceinline__ size_t physicalOffsetCUDA(size_t linear, const TensorLayout& L);
 
 __global__ void addKernel(
     const float* __restrict__ lhs, const TensorLayout lhsLayout, 
     const float* __restrict__ rhs, const TensorLayout rhsLayout, 
-    float* __restrict__ out, const TensorLayout outLayout, int count);
+    float* __restrict__ out, const TensorLayout outLayout, size_t count);
 
 __global__ void subKernel(
     const float* __restrict__ lhs, const TensorLayout lhsLayout, 
     const float* __restrict__ rhs, const TensorLayout rhsLayout, 
-    float* __restrict__ out, const TensorLayout outLayout, int count);
+    float* __restrict__ out, const TensorLayout outLayout, size_t count);
 
 __global__ void mulKernel(
     const float* __restrict__ lhs, const TensorLayout lhsLayout, 
     const float* __restrict__ rhs, const TensorLayout rhsLayout, 
-    float* __restrict__ out, const TensorLayout outLayout, int count);
+    float* __restrict__ out, const TensorLayout outLayout, size_t count);
 
 __global__ void divKernel(
     const float* __restrict__ lhs, const TensorLayout lhsLayout, 
     const float* __restrict__ rhs, const TensorLayout rhsLayout, 
-    float* __restrict__ out, const TensorLayout outLayout, int count);
+    float* __restrict__ out, const TensorLayout outLayout, size_t count);
 
 
-__global__ void fillKernel(float* __restrict__ data, float value, int count);
+__global__ void fillKernel(float* __restrict__ data, float value, size_t count);
 
 __global__ void setKernel(
     float* __restrict__ dst, const TensorLayout dstLayout,
-    const float* __restrict__ src, const TensorLayout srcLayout, int count);
+    const float* __restrict__ src, const TensorLayout srcLayout, size_t count);
 
 __global__ void checkAllEqualKernel(
     const float* __restrict__ lhs, const TensorLayout lhsLayout,
     const float* __restrict__ rhs, const TensorLayout rhsLayout,
-    int* isEqualFlag, int count);
+    int* isEqualFlag, size_t count);
 
 #endif // KERNELS_CUH
